requirements.txt
# Core OCR Dependencies
fastapi==0.111.0
uvicorn[standard]==0.30.0
surya-ocr
python-multipart
pytesseract
opencv-python-headless
Pillow
httpx

# AI Vision Providers
google-generativeai>=0.8.0  # Gemini
openai>=1.0.0  # OpenAI GPT-4 Vision
anthropic>=0.18.0  # Claude with vision

# Agent Framework (future use)
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0

# Vector Database & Memory
chromadb>=0.5.0
faiss-cpu>=1.8.0

# Testing & Metrics
pytest>=8.0.0
pytest-asyncio>=0.23.0
python-Levenshtein>=0.25.0  # For CER/WER calculations

# Additional OCR Engines
paddleocr>=2.7.0
transformers>=4.40.0  # For TrOCR and other models
torch>=2.0.0

# Dashboard & UI
jinja2>=3.1.0
websockets>=12.0

# Data & Utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# Monitoring & Logging
loguru>=0.7.0
